requirements-dev.txt
watchdog==6.0.0
ruff
icecream
django-debug-toolbar==6.0.0
django-extensions==4.1
pytest-django==4.11.1
bandit==1.9.1
sphinx==8.2.3
